+++ src/Callisto/Controls/Menu/ToggleMenuItem.cs
@@ -1,13 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿﻿//
+// Copyright (c) 2012 Tim Heuer
+//
+// Licensed under the Microsoft Public License (Ms-PL) (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://opensource.org/licenses/Ms-PL.html
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
 using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Documents;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
 
 namespace Callisto.Controls
 {
@@ -15,7 +22,7 @@ namespace Callisto.Controls
     {
         public ToggleMenuItem()
         {
-            this.DefaultStyleKey = typeof(ToggleMenuItem);
+            DefaultStyleKey = typeof(ToggleMenuItem);
         }
 
         public bool IsChecked